backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3

# PDF processing
pdfplumber==0.11.4
pdfminer.six==20231228
pdf2image==1.17.0
Pillow==11.0.0

# OCR
pytesseract==0.3.13
opencv-python==4.10.0.84

# Text processing
ftfy==6.3.1
regex==2024.11.6

# Export formats
python-docx==1.1.2
ebooklib==0.18

# Image processing for OCR
numpy==1.26.4
scikit-image==0.24.0

# Utilities
typing-extensions==4.12.2
